add tasks, rework internationalization

+++ travelcrm/static/js/jeasyui/src/jquery.form.js
@@ -1,12 +1,15 @@
+﻿/**
+ * jQuery EasyUI 1.3.6
+ * 
+ * Copyright (c) 2009-2014 www.jeasyui.com. All rights reserved.
+ *
+ * Licensed under the GPL license: http://www.gnu.org/licenses/gpl.txt
+ * To use it on other terms please contact us at info@jeasyui.com
+ *
+ */
 /**
  * form - jQuery EasyUI
  * 
- * Copyright (c) 2009-2013 www.jeasyui.com. All rights reserved.
- *
- * Licensed under the GPL or commercial licenses
- * To use it on other terms please contact us: info@jeasyui.com
- * http://www.gnu.org/licenses/gpl.txt
- * http://www.jeasyui.com/license_commercial.php
  */
 (function($){
 	/**
@@ -222,8 +225,14 @@
 				this.value = '';
 			} else if (t == 'file'){
 				var file = $(this);
-				file.after(file.clone().val(''));
-				file.remove();
+				var newfile = file.clone().val('');
+				newfile.insertAfter(file);
+				if (file.data('validatebox')){
+					file.validatebox('destroy');
+					newfile.validatebox();
+				} else {
+					file.remove();
+				}
 			} else if (t == 'checkbox' || t == 'radio'){
 				this.checked = false;
 			} else if (tag == 'select'){
@@ -334,7 +343,12 @@
 	$.fn.form.methods = {
 		submit: function(jq, options){
 			return jq.each(function(){
-				ajaxSubmit(this, $.extend({}, $.fn.form.defaults, options||{}));
+				var opts = $.extend({}, 
+					$.fn.form.defaults, 
+					$.data(this, 'form') ? $.data(this, 'form').options : {},
+					options||{}
+				);
+				ajaxSubmit(this, opts);
 			});
 		},
 		load: function(jq, data){
